feat: command palette: allow bookmark search
fix: event listener was being attached on the host site causing the host site to not function properly. solution: attach listener to the iframe dom and added a flag to differenctiate side panel and content script.

search: allow search filter option for bookmarks, and send the filter to background to search accordingly.
style: show link type indicator for bookmark and hsitory.

other fixes and improvemnts

Signed-off-by: Manish Mandal <manishmandalj@gmail.com>